requirements.txt
openpyxl>=3.0.0
Flask==2.3.3
Flask-SocketIO==5.3.6
requests==2.31.0
python-socketio==5.9.0
python-engineio==4.7.1
openpyxl==3.1.2
reportlab==4.0.4
python-dotenv==1.0.0
